+++ ReceiptME/ReceiptME/Info.plist
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CFBundleURLTypes</key>
+	<array>
+		<dict>
+			<key>CFBundleTypeRole</key>
+			<string>Editor</string>
+			<key>CFBundleURLName</key>
+			<string>GIDClientID</string>
+			<key>CFBundleURLSchemes</key>
+			<array>
+				<string>com.googleusercontent.apps.650383131525-anbr9ft0hfl03jbhbl21sokpgchc12tg</string>
+			</array>
+		</dict>
+	</array>
+	<key>GIDClientID</key>
+	<string>650383131525-anbr9ft0hfl03jbhbl21sokpgchc12tg.apps.googleusercontent.com</string>
+</dict>
+</plist>
